Clean up code.
Make the `ModelViewerState` of Web and Mobile share the same  `ModelViewer`.
