+++ lib/model_viewer_plus.dart
@@ -10,8 +10,4 @@
 ///    for the web component this widget embeds.
 library model_viewer_plus;
 
-// export 'src/model_viewer_plus.dart';
-
-export 'src/model_viewer_plus.dart'
-    if (dart.library.io) 'src/model_viewer_plus.dart'
-    if (dart.library.js) 'src/model_viewer_plus_web.dart';
+export 'src/model_viewer_plus.dart';
