+++ lib/src/model_viewer_plus_stub.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+import 'model_viewer_plus.dart';
+
+class ModelViewerState extends State<ModelViewer> {
+  @override
+  void initState() {
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    // ignore: todo
+    // TODO: implement build
+    throw UnimplementedError();
+  }
+}
